refactor: navigation 관련 로직 분리
